# openapi.yaml
openapi: 3.0.0
paths:
...
  /categories:
    get:
      description: Retrieves all product categories with their child categories
      operationId: listCategories
      parameters: []
      responses:
        '200':
          description: Categories retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryWithChildrenResponseDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: List all categories
      tags:
        - Categories
    post:
      description: Creates a new product category
      operationId: createCategory
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryDto'
      responses:
        '201':
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponseDto'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: Create category
      tags:
        - Categories
  /categories/{id}:
    put:
      description: Updates an existing product category
      operationId: updateCategory
      parameters:
        - name: id
          required: true
          in: path
          description: Category UUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryDto'
      responses:
        '200':
          description: Category updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponseDto'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: Update category
      tags:
        - Categories
    delete:
      description: Deletes a product category
      operationId: deleteCategory
      parameters:
        - name: id
          required: true
          in: path
          description: Category UUID
          schema:
            type: string
      responses:
        '200':
          description: Category deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponseDto'
        '400':
          description: Invalid UUID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: Delete category
      tags:
        - Categories
...
info:
  title: LibreStock Inventory API
  description: REST API for LibreStock Inventory Management System
  version: 1.0.0
  contact:
    name: API Support
    url: ''
    email: ''
  license:
    name: MIT
    url: ''
...
servers:
  - url: http://localhost:8080
    description: Development Server
  - url: https://api.librestock-inventory.com
    description: Production Server
components:
  securitySchemes:
    BearerAuth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Clerk JWT token
  schemas:
...
    ErrorResponseDto:
      type: object
      properties:
        error:
          type: string
          example: Resource not found
      required:
        - error
...
    CategoryWithChildrenResponseDto:
      type: object
      properties:
        created_at:
          format: date-time
          type: string
          description: Creation timestamp
        updated_at:
          format: date-time
          type: string
          description: Last update timestamp
        id:
          type: string
          description: Unique identifier
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          description: Category name
          example: Electronics
        parent_id:
          type: object
          description: Parent category ID
          format: uuid
          nullable: true
        description:
          type: object
          description: Category description
          nullable: true
          example: Electronic equipment and components
        children:
          description: Child categories
          type: array
          items:
            $ref: '#/components/schemas/CategoryWithChildrenResponseDto'
      required:
        - created_at
        - updated_at
        - id
        - name
        - parent_id
        - description
        - children
    CreateCategoryDto:
      type: object
      properties:
        name:
          type: string
          description: Category name
          example: Electronics
          minLength: 1
          maxLength: 100
        parent_id:
          type: string
          description: Parent category ID
          format: uuid
          nullable: true
        description:
          type: string
          description: Category description
          maxLength: 500
          nullable: true
      required:
        - name
    CategoryResponseDto:
      type: object
      properties:
        created_at:
          format: date-time
          type: string
          description: Creation timestamp
        updated_at:
          format: date-time
          type: string
          description: Last update timestamp
        id:
          type: string
          description: Unique identifier
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          description: Category name
          example: Electronics
        parent_id:
          type: object
          description: Parent category ID
          format: uuid
          nullable: true
        description:
          type: object
          description: Category description
          nullable: true
          example: Electronic equipment and components
      required:
        - created_at
        - updated_at
        - id
        - name
        - parent_id
        - description
    UpdateCategoryDto:
      type: object
      properties:
        name:
          type: string
          description: Category name
          minLength: 1
          maxLength: 100
        parent_id:
          type: string
          description: Parent category ID
          format: uuid
          nullable: true
        description:
          type: string
          description: Category description
          maxLength: 500
          nullable: true
    MessageResponseDto:
      type: object
      properties:
        message:
          type: string
          example: Operation completed successfully
      required:
        - message
...